+++ shared/runtime/interrupt_char.h
@@ -28,7 +28,7 @@
 
 #include "py/mpconfig.h"
 
-extern MP_IPT int mp_interrupt_char;
-void mp_hal_set_interrupt_char(int c);
+extern MP_IPT int16_t mp_interrupt_char;
+void mp_hal_set_interrupt_char(int16_t c);
 
 #endif // MICROPY_INCLUDED_LIB_UTILS_INTERRUPT_CHAR_H
